remote: extract SSH/transfer primitives into AptOfflineSSHLib
Move _run_quiet, _ssh_run, _detect_transfer_method, _transfer_get,
_transfer_put and _latest_state out of AptOfflineCoreLib into a
dedicated AptOfflineSSHLib module. AptOfflineCoreLib stays focused
on apt package management; AptOfflineRemoteLib imports each from
its proper home.
